+++ ui/components/filters/index.ts
@@ -1,3 +1,4 @@
+export * from "./clear-filters-button";
 export * from "./custom-account-selection";
 export * from "./custom-checkbox-muted-findings";
 export * from "./custom-date-picker";